Refactor to abstract out redis controller
