+++ matching_service/schemas/matching.py
@@ -2,10 +2,14 @@ from uuid import UUID
 from pydantic import BaseModel
 
 
+class MatchingCriteriaSchema(BaseModel):
+    topics: list[str]
+    difficulty: list[str]
+    primary_lang: str | None = None
+    secondary_lang: list[str] = []
+    proficiency: int = 0
+
+
 class MatchUserRequestSchema(BaseModel):
     user_id: UUID
-    topics: list[str]
-    difficulty: list = []
-    primary_lang: None = None
-    secondary_lang: list = []
-    proficiency: int = 0
+    criteria: MatchingCriteriaSchema
